requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
pymongo==3.6.1
PySocks==1.6.8
requests==2.18.4
schedule==0.5.0
urllib3==1.22